Automatically discover installed MU plugins, and autoload them
If the "mu-plugins" key is not present in "extra" in composer.json, look for folders in the mu-plugins directory, find their entrypoint PHP file, and add them to the list of plugins to autoload.

If the "mu-plugins" key is defined (whether it's empty or not), the existing behavior is preserved and that list is used as-is.